Refactor: api.js -> config.js로 변경 및 포트 번호 변경
