+++ src/component/Nav/Component/LushProducts/LushProducts.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import LushProductsList from "./LushProductsList";
 import "./LushProductsList.scss";
-import { API, CATEGORY_PATH } from "../../../../utils/api";
+import { API, CATEGORY_PATH } from "../../../../config";
 
 class LushProducts extends Component {
   constructor() {
