+++ src/config.js
@@ -1,0 +1,2 @@
+export const API = "http://10.89.6.238:8000";
+export const CATEGORY_PATH = "/category";
