--- src/utils/api.js
@@ -1,2 +1,0 @@
-export const API = "http://192.168.0.5:8000";
-export const CATEGORY_PATH = "/category";
